+++ app/src/main/java/com/example/simplemoviesapp/model/data_classes/movie_images_response/MovieImagesResponse.kt
@@ -8,7 +8,7 @@ import com.google.gson.annotations.SerializedName
 @Entity(tableName = "MovieImagesResponse", primaryKeys = ["id"])
 data class MovieImagesResponse (
 
-  @ColumnInfo(name = "id"               ) @SerializedName("id"        ) var id        : Int                  = 0,
+  @ColumnInfo(name = "id"               ) @SerializedName("id"        ) var id        : Long                 = 0,
   @ColumnInfo(name = "posters"          ) @SerializedName("posters"   ) var posters   : ArrayList<Posters>   = arrayListOf()
 
 )
